requirements.txt
Django==3.1.2
djangorestframework==3.12.1
greenlet==0.4.17
gunicorn==20.0.4
meinheld==1.0.2
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sentry-sdk==0.19.0
social-auth-app-django==4.0.0
django-cors-headers==3.5.0
social-auth-core==3.3.3
